Major revision

+++ src/attribute.jl
@@ -1,5 +1,5 @@
 mutable struct Attribute
     name::String
-    min::Float32
-    max::Float32
+    min::Float64
+    max::Float64
 end
